remove package-lock.json
package-lock.json is for top-level apps, not for packages. If you use it inside a package it will only confuse you when you run npm install and get different (locked) dependencies than your users get when they run npm install package-name.

One possible advantage of package-lock.json for modules is that it documents which set of dependencies we know to work (in case a dependency suddenly causes our tests to start failing). However, I haven't found this useful in practice.